fix(MissionsList): Moving Modal into its own file
